+++ packages/atomic/cypress/integration/facets/category-facet/category-facet-assertions.ts
@@ -4,10 +4,7 @@ import {
 } from '../../../utils/componentUtils';
 import {RouteAlias} from '../../../utils/setupComponent';
 import {ResultListSelectors} from '../../result-list/result-list-selectors';
-import {
-  CategoryFacetSelectors,
-  BreadcrumbSelectors,
-} from './category-facet-selectors';
+import {CategoryFacetSelectors} from './category-facet-selectors';
 import {hierarchicalField} from './category-facet-actions';
 import {should} from '../../common-assertions';
 
@@ -39,16 +36,6 @@ export function assertNumberOfParentValues(value: number) {
   });
 }
 
-export function assertPathInBreadcrumb(path: string[]) {
-  const ellipsedPath =
-    path.length > 3 ? path.slice(0, 1).concat(['...'], path.slice(-2)) : path;
-  const joinedPath = ellipsedPath.join(' / ');
-  // TODO: fix breadcrumb tests
-  it.skip(`should display the selected path "${joinedPath}" in the breadcrumbs`, () => {
-    BreadcrumbSelectors.breadcrumbButton().first().contains(joinedPath);
-  });
-}
-
 export function assertPathInUrl(path: string[]) {
   const categoryFacetListInUrl = path.join(',');
   it(`should display the selected path "${categoryFacetListInUrl}" in the url`, () => {
@@ -56,13 +43,6 @@ export function assertPathInUrl(path: string[]) {
       categoryFacetListInUrl
     )}`;
     cy.url().should('include', urlHash);
-  });
-}
-
-export function assertNoBreadcrumb() {
-  // TODO: fix breadcrumb tests
-  it.skip('should not have any breadcrumb', () => {
-    BreadcrumbSelectors.breadcrumbButton().should('not.exist');
   });
 }