+++ packages/atomic/cypress/integration/result-list/result-components/result-badge.cypress.ts
@@ -82,6 +82,10 @@ describe('Result Badge Component', () => {
         it('renders the localized text', () => {
           ResultBadgeSelectors.resultText().should('have.text', localizedText);
         });
+
+        CommonAssertions.assertAccessibility(
+          ResultBadgeSelectors.firstInResult
+        );
       });
 
       describe('when a label is specified', () => {